requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-dotenv
passlib[bcrypt]
python-jose
pydantic
